http request to api with token headers in role

+++ src/app/role/ListarRoles.js
@@ -3,6 +3,9 @@ import {Link, Redirect} from 'react-router-dom';
 import { Button, Card, CardBody, CardHeader, Col, Row, Table } from 'reactstrap';
 import API from '../../services/api';
 
+import AuthService from '../../services/AuthService';
+const auth = new AuthService();
+
 export default class ListarRoles extends React.Component{
   constructor(props){
     super(props);
@@ -10,27 +13,29 @@ export default class ListarRoles extends React.Component{
       roles: [],
     }
 
-    this.deteleRole = this.deteleRole.bind(this);
+    this.deleteRole = this.deleteRole.bind(this);
   }
 
-  deteleRole = (id) => {
-    API.delete(`/role/${id}`)
+  componentDidMount(){
+    API.get(`/role`, { headers: {"Authorization" : `Bearer ${auth.getToken()}`} })
+      .then(res => {
+        //console.log(res.data);
+        const r = res.data;
+        this.setState({ roles: r });
+      })
+      .catch(err => {
+        console.log(err);
+      });;
+  }
+
+  deleteRole = (id) => {
+    API.delete(`/role/${id}`, { headers: {"Authorization" : `Bearer ${auth.getToken()}`} })
       .then(res => {
         const arr = this.state.roles.filter(r => r.id !== id);
         this.setState({roles: arr});
       })
       .catch(err => {
-        //FIXME
-        alert("soy un error" + err);
-      });
-  }
-
-  componentDidMount(){
-    API.get(`/role`)
-      .then(res => {
-        //console.log(res.data);
-        const r = res.data;
-        this.setState({ roles: r });
+        console.log(err);
       });
   }
 
@@ -60,7 +65,7 @@ export default class ListarRoles extends React.Component{
                         &nbsp;
                         <Link to={`/roles/edit/${role.id}`}><Button size="sm" color="success"><i className="icon-pencil"></i></Button></Link>
                         &nbsp;
-                        <Button onClick={ () => this.deteleRole(role.id) } size="sm" color="danger"><i className="fa fa-trash"></i></Button>
+                        <Button onClick={ () => this.deleteRole(role.id) } size="sm" color="danger"><i className="fa fa-trash"></i></Button>
                       </td>
                     </tr>
                   )}
